requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
base58==2.1.1
bitarray==2.5.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Cython==0.29.30
cytoolz==0.11.2
dateparser==1.0.0
distlib==0.3.4
dydx-v3-python==1.6.0
ecdsa==0.16.0
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
filelock==3.7.1
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
importlib-metadata==0.23
importlib-resources==5.7.1
ipfshttpclient==0.8.0a2
jsonschema==4.6.0
lru-dict==1.1.7
lxml==4.9.0
more-itertools==8.13.0
mpmath==1.0.0
multiaddr==0.0.9
multidict==6.0.2
multitasking==0.0.10
netaddr==0.8.0
numpy==1.21.6
packaging==21.3
pandas==1.3.5
parsimonious==0.8.1
platformdirs==2.5.2
pluggy==0.13.1
protobuf==3.20.1
py==1.11.0
pycryptodome==3.14.1
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==4.6.11
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.6.2
requests==2.27.1
requests-mock==1.6.0
rlp==2.0.1
six==1.16.0
sympy==1.6
toml==0.10.2
toolz==0.11.2
tox==3.13.2
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
varint==1.0.2
virtualenv==20.14.1
wcwidth==0.2.5
web3==5.29.2
websockets==9.1
yarl==1.7.2
yfinance==0.1.70
zipp==3.8.0
